broken: Make initial changes for ONNX export
The goal is to get DLRM as ONNX. Not sure yet whether it's possible, but
that's why I'm trying.

Modify the DLRMPredictFactory constructor to optionally accept a
DLRMModelConfig. Copy the packager to a new file and try to instantiate
a DLRMPredictModule through the DLRMPredictFactory before exporting it
to ONNX.

This code fails with a TypeError related to abstract class
instantiation.