Implement unsharded sparse arch and DLRM
The unsharded sparse arch is accepts a list of JaggedTensors instead of
a KeyedJaggedTensor. It also explicitly manages a list of EmbeddingBags
rather than an EmbeddingBagCollections. It is named SparseArchUnsharded.

The unsharded DLRM handles the above sparse arch by not using
KeyedJaggedTensors. It is named DLRM_DCN_Unsharded.

Instantiation of the class fails with an error citing autograd, fbgemm,
and JaggedTensor. Not sure what the issue is yet. It may be a simple
package requirement.